Nix tree.orm.DocumentSet from most (all) places?

+++ app/views/DocumentSet/showHtml.json.scala
@@ -5,10 +5,10 @@ import play.api.libs.json.{Json,JsValue}
 import play.api.mvc.RequestHeader
 
 import models.User
-import org.overviewproject.tree.orm.DocumentSet
+import org.overviewproject.models.DocumentSet
 
 object showHtml {
-  def apply(user: User,documentSet: DocumentSet, nViews: Long)(implicit lang: Lang, request: RequestHeader): JsValue = {
+  def apply(user: User, documentSet: DocumentSet, nViews: Int)(implicit lang: Lang, request: RequestHeader): JsValue = {
     Json.obj(
       "id" -> documentSet.id,
       "html" -> views.html.DocumentSet._documentSet(documentSet, nViews, user).toString
